+++ index.html
@@ -8,10 +8,12 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
   <script src="functions.js"></script>
-  <script srs="main.js"></script>
+
+  <title>Модульная структура</title>
 </head>
 
 <body>
+  <script type='module' srs="main.js"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
